[HER-105]Follow-up Move MPagination and PPagination
