+++ src/Business/Business.csproj
@@ -8,7 +8,6 @@
 
     <ItemGroup>
       <Folder Include="Camera\Common" />
-      <Folder Include="Pagination\Common" />
     </ItemGroup>
 
 </Project>
