+++ src/Business/Pagination/ICreatePagination.cs
@@ -1,6 +1,7 @@
 namespace Business.Pagination
 {
     using System.Linq;
+    using Business.Pagination.Models;
 
     public interface ICreatePagination<TEntity> : ICommand<IQueryable<TEntity>, MPagination<TEntity>>
     {
