+++ src/Business/Pagination/Models/PPagination.cs
@@ -1,4 +1,4 @@
-namespace Business.Pagination
+namespace Business.Pagination.Models
 {
     public class PPagination
     {
